Fix cancellation of TransactionBuilder::start
